Extract scan-native composition hook from useAppShellComposition
